fix for cr suggestions
