+++ apps/dialtone-documentation/docs/.vuepress/baseComponents/TokenTable.vue
@@ -79,26 +79,7 @@
 <script>
 import CopyButton from './CopyButton.vue';
 import TokenExample from './TokenExample.vue';
-
-export const FORMAT_MAP = {
-  CSS: 'css/variables',
-  Android: 'compose/object',
-  iOS: 'ios-swift/enum.swift',
-};
-
-export const THEMES = [
-  { value: 'light', label: 'Light' },
-  { value: 'dark', label: 'Dark' },
-];
-
-export const CATEGORY_MAP = {
-  color: ['color', 'opacity', 'theme'],
-  typography: ['typography', 'font'],
-  size: ['size'],
-  space: ['space'],
-  shadow: ['shadow'],
-  component: ['avatar', 'badge', 'checkbox', 'icon', 'inputs', 'action'],
-};
+import { CATEGORY_MAP } from '../common/constants';
 
 export default {
   name: 'TokenTable',
