+++ apps/dialtone-documentation/docs/tokens/index.md
@@ -5,7 +5,7 @@ description: Multi-platform values that make up Dialtone's design language
 
 Design tokens are all the defined values that are part of the design system, such as color, spacing, typography, and more. Design tokens are published in many different forms so they can be consumed on all platforms (CSS, LESS, Android, iOS, etc).
 
-See the [dialtone-tokens repository](https://github.com/dialpad/dialtone-tokens) for usage info.
+See the [dialtone tokens README](https://github.com/dialpad/dialtone/tree/staging/packages/dialtone-tokens#readme) for usage info.
 
 <all-tokens />
